+++ CHANGELOG.rst
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
+2.129.2 - 2024-07-09
+====================
+
+Added
+-----
+* Support for cost management of shared resources in the Usage service
+* Support for creating new databases with oracle key vault on Exadata Cloud at Customer in the Database service
+* Support for confirming key store details on Exadata Cloud at Customer in the Database service
+
+====================
 2.129.1 - 2024-07-02
 ====================